+++ .github/workflows/test-tools.yml
@@ -46,11 +46,5 @@ jobs:
       - name: Lint
         run: npm run lint
 
-      - name: Test Spec Merger
-        run: npm run test -- tools/tests/merger
-
-      - name: Test Spec Linter
-        run: npm run test -- tools/tests/linter
-
-      - name: Test Spec Tester
-        run: npm run test -- tools/tests/tester --runInBand
+      - name: Tests
+        run: npm run test
