+++ package.json
@@ -11,7 +11,9 @@
     "lint": "eslint .",
     "lint--fix": "eslint . --fix",
     "merge": "ts-node tools/src/merger/merge.ts",
-    "test": "jest",
+    "test": "npm run test:unit && npm run test:integ",
+    "test:unit": "jest --testMatch='**/*.test.ts' --testPathIgnorePatterns=/integ/",
+    "test:integ": "jest --testMatch='**/integ/*.test.ts' --runInBand",
     "test:spec": "ts-node tools/src/tester/start.ts"
   },
   "dependencies": {
